modify: eb_newline to actually increase the vec when newline
This is important because eb_backspace remove lines by the index
of vector. Although there is pro for lazy allocation for actual line
buffer, no-allocating line vector itself increase ambiguty too much.

Also this commit removes the feature that eb_get_line returns NULL
when its index is not correct, because this is ambiguouse with actual
line pointer is NULL on line vector and causes unfindable bug.

[minor]
add assert on line_split