fix: SIGBUS when eb_save_file
When eb_save_file is called, it opens the file with
fopen(eb->file_name, "w").  Since opening with "w" mode truncates the
file size to 0, accessing the memory region previously mmap-ed to that
file results in a SIGBUS error.  To prevent this, the content is now
first written to a temporary file at /tmp/%s.tmpXXXXXX, and then
renamed to eb->file_name.